split rules into separate files
