+++ tests/js/spec/views/settings/projectFiltersAndSampling/errors.spec.tsx
@@ -1,0 +1,483 @@
+import {
+  fireEvent,
+  screen,
+  waitForElementToBeRemoved,
+  within,
+} from 'sentry-test/reactTestingLibrary';
+
+import {commonConditionCategories, renderComponent, renderModal} from './utils';
+
+describe('Filters and Sampling - Error rule', function () {
+  MockApiClient.addMockResponse({
+    url: '/projects/org-slug/project-slug/',
+    method: 'GET',
+    body: TestStubs.Project(),
+  });
+
+  it('edit rule', async function () {
+    MockApiClient.addMockResponse({
+      url: '/projects/org-slug/project-slug/',
+      method: 'GET',
+      body: TestStubs.Project({
+        dynamicSampling: {
+          rules: [
+            {
+              sampleRate: 0.1,
+              type: 'error',
+              condition: {
+                op: 'and',
+                inner: [
+                  {
+                    op: 'glob',
+                    name: 'event.release',
+                    value: ['1*'],
+                  },
+                ],
+              },
+              id: 39,
+            },
+            {
+              sampleRate: 0.2,
+              type: 'trace',
+              condition: {
+                op: 'and',
+                inner: [
+                  {
+                    op: 'glob',
+                    name: 'trace.release',
+                    value: ['1.2.3'],
+                  },
+                ],
+              },
+              id: 40,
+            },
+          ],
+          next_id: 43,
+        },
+      }),
+    });
+
+    MockApiClient.addMockResponse({
+      url: '/projects/org-slug/project-slug/',
+      method: 'PUT',
+      body: TestStubs.Project({
+        dynamicSampling: {
+          rules: [
+            {
+              sampleRate: 0.5,
+              type: 'error',
+              condition: {
+                op: 'and',
+                inner: [
+                  {
+                    op: 'glob',
+                    name: 'event.release',
+                    value: ['[I3].[0-9]'],
+                  },
+                ],
+              },
+              id: 44,
+            },
+            {
+              sampleRate: 0.2,
+              type: 'trace',
+              condition: {
+                op: 'and',
+                inner: [
+                  {
+                    op: 'glob',
+                    name: 'trace.release',
+                    value: ['1.2.3'],
+                  },
+                ],
+              },
+              id: 40,
+            },
+          ],
+          next_id: 43,
+        },
+      }),
+    });
+
+    MockApiClient.addMockResponse({
+      url: '/organizations/org-slug/tags/release/values/',
+      method: 'GET',
+      body: [{value: '[I3].[0-9]'}],
+    });
+
+    renderComponent();
+
+    // Error rules container
+    expect(
+      screen.queryByText('There are no error rules to display')
+    ).not.toBeInTheDocument();
+    expect(screen.getByText('Errors only')).toBeInTheDocument();
+
+    // Transaction traces and individual transactions rules container
+    expect(
+      screen.queryByText('There are no transaction rules to display')
+    ).not.toBeInTheDocument();
+    const transactionTraceRules = screen.getByText('Transaction traces');
+    expect(transactionTraceRules).toBeInTheDocument();
+
+    const editRuleButtons = screen.getAllByLabelText('Edit Rule');
+    expect(editRuleButtons).toHaveLength(2);
+
+    // Open rule modal - edit error rule
+    await renderModal(editRuleButtons[0]);
+
+    // Modal content
+    expect(screen.getByText('Edit Error Sampling Rule')).toBeInTheDocument();
+    expect(screen.queryByText('Tracing')).not.toBeInTheDocument();
+
+    // Release Field
+    expect(await screen.findByTestId('autocomplete-release')).toBeInTheDocument();
+
+    // Release field is not empty
+    const releaseFieldValues = screen.getByTestId('multivalue');
+    expect(releaseFieldValues).toBeInTheDocument();
+    expect(releaseFieldValues).toHaveTextContent('1*');
+
+    // Button is enabled - meaning the form is valid
+    const saveRuleButton = screen.getByRole('button', {name: 'Save Rule'});
+    expect(saveRuleButton).toBeInTheDocument();
+    expect(saveRuleButton).toBeEnabled();
+
+    // Sample rate field
+    const sampleRateField = screen.getByPlaceholderText('\u0025');
+    expect(sampleRateField).toBeInTheDocument();
+
+    // Sample rate is not empty
+    expect(sampleRateField).toHaveValue(10);
+
+    // Clear release field
+    fireEvent.keyDown(screen.getByLabelText('Search or add a release'), {
+      key: 'Backspace',
+    });
+
+    // Release field is now empty
+    expect(screen.queryByTestId('multivalue')).not.toBeInTheDocument();
+
+    expect(screen.getByRole('button', {name: 'Save Rule'})).toBeDisabled();
+
+    // Type into realease field
+    fireEvent.change(screen.getByLabelText('Search or add a release'), {
+      target: {value: '[I3].[0-9]'},
+    });
+
+    // Autocomplete suggests options
+    const autocompleteOptions = screen.getByTestId('option');
+    expect(autocompleteOptions).toBeInTheDocument();
+    expect(autocompleteOptions).toHaveTextContent('[I3].[0-9]');
+
+    // Click on the suggested option
+    fireEvent.click(autocompleteOptions);
+
+    expect(screen.getByRole('button', {name: 'Save Rule'})).toBeEnabled();
+
+    // Clear sample rate field
+    fireEvent.change(sampleRateField, {target: {value: null}});
+
+    expect(screen.getByRole('button', {name: 'Save Rule'})).toBeDisabled();
+
+    // Update sample rate field
+    fireEvent.change(sampleRateField, {target: {value: 50}});
+
+    // Save button is now enabled
+    const saveRuleButtonEnabled = screen.getByRole('button', {name: 'Save Rule'});
+    expect(saveRuleButtonEnabled).toBeEnabled();
+
+    // Click on save button
+    fireEvent.click(saveRuleButtonEnabled);
+
+    // Modal will close
+    await waitForElementToBeRemoved(() => screen.getByText('Edit Error Sampling Rule'));
+
+    // Error rules panel is updated
+    expect(screen.getByText('Errors only')).toBeInTheDocument();
+
+    expect(screen.getAllByText('Release')).toHaveLength(2);
+
+    // Old values
+    expect(screen.queryByText('1*')).not.toBeInTheDocument();
+    expect(screen.queryByText('10%')).not.toBeInTheDocument();
+
+    // New values
+    expect(screen.getByText('[I3].[0-9]')).toBeInTheDocument();
+    expect(screen.getByText('50%')).toBeInTheDocument();
+  });
+
+  it('delete rule', async function () {
+    MockApiClient.addMockResponse({
+      url: '/projects/org-slug/project-slug/',
+      method: 'GET',
+      body: TestStubs.Project({
+        dynamicSampling: {
+          rules: [
+            {
+              sampleRate: 0.2,
+              type: 'error',
+              condition: {
+                op: 'and',
+                inner: [
+                  {
+                    op: 'glob',
+                    name: 'event.release',
+                    value: ['1.2.3'],
+                  },
+                ],
+              },
+              id: 39,
+            },
+            {
+              sampleRate: 0.2,
+              type: 'trace',
+              condition: {
+                op: 'and',
+                inner: [
+                  {
+                    op: 'glob',
+                    name: 'trace.release',
+                    value: ['1.2.3'],
+                  },
+                ],
+              },
+              id: 40,
+            },
+          ],
+          next_id: 43,
+        },
+      }),
+    });
+
+    MockApiClient.addMockResponse({
+      url: '/projects/org-slug/project-slug/',
+      method: 'PUT',
+      body: TestStubs.Project({
+        dynamicSampling: {
+          rules: [
+            {
+              sampleRate: 0.2,
+              type: 'trace',
+              condition: {
+                op: 'and',
+                inner: [
+                  {
+                    op: 'glob',
+                    name: 'trace.release',
+                    value: ['1.2.3'],
+                  },
+                ],
+              },
+              id: 40,
+            },
+          ],
+          next_id: 43,
+        },
+      }),
+    });
+
+    renderComponent();
+
+    // Error rules container
+    expect(
+      screen.queryByText('There are no error rules to display')
+    ).not.toBeInTheDocument();
+    expect(screen.getByText('Errors only')).toBeInTheDocument();
+
+    // Transaction traces and individual transactions rules container
+    expect(
+      screen.queryByText('There are no transaction rules to display')
+    ).not.toBeInTheDocument();
+    const transactionTraceRules = screen.getByText('Transaction traces');
+    expect(transactionTraceRules).toBeInTheDocument();
+
+    const deleteRuleButtons = screen.getAllByLabelText('Delete Rule');
+    expect(deleteRuleButtons).toHaveLength(2);
+
+    // Open deletion confirmation modal - delete error rule
+    await renderModal(deleteRuleButtons[0]);
+
+    expect(
+      screen.getByText('Are you sure you wish to delete this dynamic sampling rule?')
+    ).toBeInTheDocument();
+
+    expect(screen.getByText('Confirm')).toBeInTheDocument();
+
+    // Confirm deletion
+    fireEvent.click(screen.getByText('Confirm'));
+
+    // Confirmation modal will close
+    await waitForElementToBeRemoved(() =>
+      screen.getByText('Are you sure you wish to delete this dynamic sampling rule?')
+    );
+
+    // Error rules panel is updated
+    expect(screen.getByText('There are no error rules to display')).toBeInTheDocument();
+
+    // There is still one transaction rule
+    expect(transactionTraceRules).toBeInTheDocument();
+  });
+
+  describe('modal', function () {
+    it('renders modal', async function () {
+      renderComponent();
+
+      // Open Modal
+      await renderModal(screen.getByText('Add error rule'), true);
+
+      // Modal content
+      expect(screen.getByText('Add Error Sampling Rule')).toBeInTheDocument();
+      expect(screen.queryByText('Tracing')).not.toBeInTheDocument();
+      expect(
+        within(screen.getByRole('dialog')).getByText('Conditions')
+      ).toBeInTheDocument();
+      expect(screen.getByText('Add Condition')).toBeInTheDocument();
+      expect(screen.getByText('Apply sampling rate to all errors')).toBeInTheDocument();
+      expect(screen.getByText('Sampling Rate \u0025')).toBeInTheDocument();
+      expect(screen.getByPlaceholderText('\u0025')).toHaveValue(null);
+      expect(screen.getByRole('button', {name: 'Cancel'})).toBeInTheDocument();
+      const saveRuleButton = screen.getByRole('button', {name: 'Save Rule'});
+      expect(saveRuleButton).toBeInTheDocument();
+      expect(saveRuleButton).toBeDisabled();
+
+      // Close Modal
+      fireEvent.click(screen.getByLabelText('Close Modal'));
+      await waitForElementToBeRemoved(() => screen.getByText('Add Error Sampling Rule'));
+    });
+
+    it('condition options', async function () {
+      renderComponent();
+
+      // Open Modal
+      await renderModal(screen.getByText('Add error rule'));
+
+      // Click on 'Add condition'
+      fireEvent.click(screen.getByText('Add Condition'));
+
+      // Autocomplete
+      expect(await screen.findByTestId('autocomplete-list')).toBeInTheDocument();
+
+      // Condition Options
+      const conditionOptions = within(
+        await screen.findByTestId('autocomplete-list')
+      ).getAllByRole('presentation');
+
+      expect(conditionOptions).toHaveLength(commonConditionCategories.length);
+
+      for (const conditionOptionIndex in conditionOptions) {
+        expect(conditionOptions[conditionOptionIndex]).toHaveTextContent(
+          commonConditionCategories[conditionOptionIndex]
+        );
+      }
+
+      // Close Modal
+      fireEvent.click(screen.getByLabelText('Close Modal'));
+      await waitForElementToBeRemoved(() => screen.getByText('Add Error Sampling Rule'));
+    });
+
+    it('save rule', async function () {
+      MockApiClient.addMockResponse({
+        url: '/projects/org-slug/project-slug/',
+        method: 'PUT',
+        body: TestStubs.Project({
+          dynamicSampling: {
+            rules: [
+              {
+                sampleRate: 0.2,
+                type: 'error',
+                condition: {
+                  op: 'and',
+                  inner: [
+                    {
+                      op: 'glob',
+                      name: 'event.release',
+                      value: ['1.2.3'],
+                    },
+                  ],
+                },
+                id: 39,
+              },
+            ],
+            next_id: 40,
+          },
+        }),
+      });
+
+      MockApiClient.addMockResponse({
+        url: '/organizations/org-slug/tags/release/values/',
+        method: 'GET',
+        body: [{value: '1.2.3'}],
+      });
+
+      renderComponent();
+
+      // Open Modal
+      await renderModal(screen.getByText('Add error rule'));
+
+      // Click on 'Add condition'
+      fireEvent.click(screen.getByText('Add Condition'));
+
+      // Autocomplete
+      const autoCompleteList = await screen.findByTestId('autocomplete-list');
+      expect(autoCompleteList).toBeInTheDocument();
+
+      // Condition Options
+      const conditionOptions = within(autoCompleteList).getAllByRole('presentation');
+
+      // Click on the first condition option
+      fireEvent.click(conditionOptions[0]);
+
+      // Release Field
+      await screen.findByTestId('autocomplete-release');
+      const releaseField = screen.getByTestId('autocomplete-release');
+      expect(releaseField).toBeInTheDocument();
+
+      // Release field is empty
+      const releaseFieldValues = screen.queryByTestId('multivalue');
+      expect(releaseFieldValues).not.toBeInTheDocument();
+
+      // Type into realease field
+      fireEvent.change(screen.getByLabelText('Search or add a release'), {
+        target: {value: '1.2.3'},
+      });
+
+      // Autocomplete suggests options
+      const autocompleteOptions = screen.getByTestId('option');
+      expect(autocompleteOptions).toBeInTheDocument();
+      expect(autocompleteOptions).toHaveTextContent('1.2.3');
+
+      // Click on the suggested option
+      fireEvent.click(autocompleteOptions);
+
+      // Button is still disabled
+      const saveRuleButton = screen.getByRole('button', {name: 'Save Rule'});
+      expect(saveRuleButton).toBeInTheDocument();
+      expect(saveRuleButton).toBeDisabled();
+
+      // Fill sample rate field
+      const sampleRateField = screen.getByPlaceholderText('\u0025');
+      expect(sampleRateField).toBeInTheDocument();
+      fireEvent.change(sampleRateField, {target: {value: 20}});
+
+      // Save button is now enabled
+      const saveRuleButtonEnabled = screen.getByRole('button', {name: 'Save Rule'});
+      expect(saveRuleButtonEnabled).toBeEnabled();
+
+      // Click on save button
+      fireEvent.click(saveRuleButtonEnabled);
+
+      // Modal will close
+      await waitForElementToBeRemoved(() => screen.getByText('Add Error Sampling Rule'));
+
+      // Error rules panel is updated
+      expect(
+        screen.queryByText('There are no error rules to display')
+      ).not.toBeInTheDocument();
+      expect(screen.getByText('Errors only')).toBeInTheDocument();
+      expect(screen.getByText('Release')).toBeInTheDocument();
+      expect(screen.getByText('1.2.3')).toBeInTheDocument();
+      expect(screen.getByText('20%')).toBeInTheDocument();
+    });
+  });
+});
